Move pathlib_map,patch_volume, and translate_settings_paths
